Updated Auction House Data Fetch + Fixed User Type Crash

+++ app/src/main/java/com/example/auctionhouseapp/User.kt
@@ -28,7 +28,7 @@ open class User {
     fun SetData(Data: MutableMap<String,Any>?){
         Name = Data!![Constants.USER_NAME] as String
         Email = Data[Constants.USER_EMAIL] as String
-        Type = UserType.getByValue( Data[Constants.USER_TYPE] as Int)
+        Type = UserType.getByValue( (Data[Constants.USER_TYPE] as Long).toInt())
         PhoneNumber = Data[Constants.USER_PHONE] as String
         Location = Data[Constants.USER_ADDR] as String
         UserID = Data[Constants.USERID] as String
